Unbinding ELF image from RouterBOOT version.
If the version does not match the one we figure on, try to work without calling
the RouterBOOT functions - using only our own. It is slower (because DCash is disabled).
But it makes possible not to depend on RouterBOOT for ELF load at all.

It is so important for ELF, because it can be launched from a backup RouterBOOT
(when the reset button is pressed and the power is turned on).